# api/openapi.yaml
openapi: 3.1.0
info:
  title: Royal Flush API
  version: 0.1.0
  description: 熟人娱乐积分德州扑克的玩家端与运营端 HTTP API。
servers:
  - url: /api/v1
paths:
...
  /auth/otp/request:
    post:
      operationId: requestOtp
      description: 仅用于本地开发兼容；生产账号使用手机号与密码。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [phone]
              properties:
                phone: { type: string, pattern: '^1[3-9][0-9]{9}$' }
      responses:
        '200':
          description: 验证码已发送
          content:
            application/json:
              schema:
                type: object
                required: [expiresAt, expiresIn]
                properties:
                  expiresAt: { type: string, format: date-time }
                  expiresIn: { type: integer, const: 300 }
                  devCode: { type: string, pattern: '^[0-9]{6}$' }
        '503': { $ref: '#/components/responses/Problem' }
  /auth/otp/verify:
    post:
      operationId: verifyOtp
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [phone, code]
              properties:
                phone: { type: string }
                code: { type: string, pattern: '^[0-9]{6}$' }
                nickname: { type: string, maxLength: 20 }
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                required: [user, balance]
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  balance: { type: integer, format: int64 }
  /auth/register:
    post:
      operationId: registerAccount
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [phone, password, nickname]
              properties:
                phone: { type: string, pattern: '^1[3-9][0-9]{9}$' }
                password: { type: string, minLength: 8, maxLength: 72 }
                nickname: { type: string, minLength: 1, maxLength: 20 }
      responses:
        '201':
          description: 注册成功并建立持久会话
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /auth/login:
    post:
      operationId: loginAccount
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [phone, password]
              properties:
                phone: { type: string, pattern: '^1[3-9][0-9]{9}$' }
                password: { type: string, minLength: 1, maxLength: 72 }
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
        '401': { $ref: '#/components/responses/Problem' }
...
  /auth/password/login:
    post:
      operationId: loginWithPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [account, password]
              properties:
                account: { type: string, minLength: 1, maxLength: 64 }
                password: { type: string, minLength: 1, maxLength: 128 }
      responses:
        '200':
          description: 管理员登录成功
          content:
            application/json:
              schema:
                type: object
                required: [user, balance]
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  balance: { type: integer, format: int64 }
  /me:
    get:
      operationId: getMe
      responses:
        '200':
          description: 当前账号
          content:
            application/json:
              schema:
                type: object
                required: [user, balance, activeRoomId]
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  balance: { type: integer, format: int64 }
                  activeRoomId: { type: string }
    patch:
      operationId: updateMe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [nickname]
              properties:
                nickname: { type: string, minLength: 1, maxLength: 20 }
      responses:
        '200':
          description: 账号资料已更新
          content:
            application/json:
              schema:
                type: object
                required: [user, balance, activeRoomId]
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  balance: { type: integer, format: int64 }
                  activeRoomId: { type: string }
        '400': { $ref: '#/components/responses/Problem' }
  /me/score-additions:
    post:
      operationId: addScore
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScoreAdditionRequest' }
      responses:
        '200':
          description: 增加成功或返回同一幂等请求的原结果
          content:
            application/json:
              schema:
                type: object
                required: [balance, entry]
                properties:
                  balance: { type: integer, format: int64 }
                  entry: { $ref: '#/components/schemas/ScoreLedgerEntry' }
        '429': { $ref: '#/components/responses/Problem' }
...
  /rooms:
    post:
      operationId: createRoom
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoomConfig' }
      responses:
        '201':
          description: 房间创建成功，房主自动坐入 0 号位
          content:
            application/json:
              schema:
                type: object
                required: [id, code, config, snapshot]
                properties:
                  id: { type: string }
                  code: { type: string }
                  config: { $ref: '#/components/schemas/RoomConfig' }
                  snapshot: { $ref: '#/components/schemas/TableSnapshot' }
  /rooms/{roomID}/public:
    parameters:
      - $ref: '#/components/parameters/RoomID'
    get:
      operationId: getPublicRoom
      security: []
      responses:
        '200':
          description: 邀请登录页可见的房间摘要
          content:
            application/json:
              schema:
                type: object
                required: [id, code, name, ownerId, ownerName, onlinePlayers, maxPlayers, config, occupiedSeats]
                properties:
                  id: { type: string }
                  code: { type: string }
                  name: { type: string }
                  ownerId: { type: string }
                  ownerName: { type: string }
                  onlinePlayers: { type: integer }
                  maxPlayers: { type: integer }
                  config: { $ref: '#/components/schemas/RoomConfig' }
                  occupiedSeats:
                    type: array
                    items: { type: integer, minimum: 0, maximum: 7 }
  /rooms/{roomID}/join:
    parameters:
      - $ref: '#/components/parameters/RoomID'
    post:
      operationId: joinRoom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [seat]
              properties:
                seat: { type: integer, minimum: 0, maximum: 7 }
      responses:
        '200':
          description: 入座成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TableSnapshot' }
        '409': { $ref: '#/components/responses/Problem' }
  /rooms/{roomID}/snapshot:
    parameters:
      - $ref: '#/components/parameters/RoomID'
    get:
      operationId: getRoomSnapshot
      responses:
        '200':
          description: 仅含当前用户底牌的权威快照
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TableSnapshot' }
  /rooms/{roomID}/commands:
    parameters:
      - $ref: '#/components/parameters/RoomID'
    post:
      operationId: sendRoomCommand
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoomCommand' }
      responses:
        '200':
          description: 命令处理结果
          content:
            application/json:
              schema:
                type: object
                required: [event, duplicate]
                properties:
                  event: { $ref: '#/components/schemas/RoomEvent' }
                  duplicate: { type: boolean }
        '409': { $ref: '#/components/responses/Problem' }
  /rooms/{roomID}/voice-token:
    parameters:
      - $ref: '#/components/parameters/RoomID'
    post:
      operationId: createVoiceToken
      responses:
        '200':
          description: LiveKit 凭证或浏览器 WebRTC 降级配置
          content:
            application/json:
              schema:
                type: object
                required: [enabled]
                properties:
                  enabled: { type: boolean }
                  transport: { type: string, enum: [livekit, webrtc] }
                  url: { type: string }
                  accessToken: { type: string }
                  iceServers:
                    type: array
                    items: { $ref: '#/components/schemas/ICEServer' }
                  expiresAt: { type: string, format: date-time }
                  reason: { type: string }
  /rooms/{roomID}/voice-events:
    parameters:
      - $ref: '#/components/parameters/RoomID'
    get:
      operationId: connectVoiceEvents
      description: 将已登录且已入座的用户升级到 WebSocket，转发 WebRTC SDP 与 ICE 信令。
      responses:
        '101':
          description: WebSocket 已建立
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /reports:
    post:
      operationId: createReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [category, detail, requestId]
              properties:
                roomId: { type: string }
                subjectUserId: { type: string }
                category: { type: string, enum: [conduct, voice, technical, other] }
                detail: { type: string, minLength: 2, maxLength: 1000 }
                requestId: { type: string, minLength: 1 }
      responses:
        '201':
          description: 举报已登记
          content:
            application/json:
              schema:
                type: object
                required: [report, duplicate]
                properties:
                  report: { $ref: '#/components/schemas/Report' }
                  duplicate: { type: boolean }
  /admin/users:
    get:
      operationId: listAdminUsers
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500 } }
      responses:
        '200':
          description: 用户与积分列表
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users: { type: array, items: { $ref: '#/components/schemas/OperationsUser' } }
  /admin/users/{userID}/ban-actions:
    parameters:
      - name: userID
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: setAdminUserBanned
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [banned, reason, requestId]
              properties:
                banned: { type: boolean }
                reason: { type: string, minLength: 1 }
                requestId: { type: string, minLength: 1 }
      responses:
        '200':
          description: 用户封禁状态
          content:
            application/json:
              schema:
                type: object
                required: [user, duplicate]
                properties:
                  user: { $ref: '#/components/schemas/OperationsUser' }
                  duplicate: { type: boolean }
...
  /admin/rooms/{roomID}:
    parameters:
      - $ref: '#/components/parameters/RoomID'
    get:
      operationId: getAdminRoom
      responses:
        '200':
          description: 不包含任何玩家底牌的运营牌桌快照
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TableSnapshot' }
  /admin/reports:
    get:
      operationId: listAdminReports
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [open, reviewing, resolved, dismissed] } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500 } }
      responses:
        '200':
          description: 举报处理队列
          content:
            application/json:
              schema:
                type: object
                required: [reports]
                properties:
                  reports: { type: array, items: { $ref: '#/components/schemas/Report' } }
  /admin/reports/{reportID}/resolution:
    parameters:
      - name: reportID
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: resolveAdminReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [status, reason, requestId]
              properties:
                status: { type: string, enum: [resolved, dismissed] }
                reason: { type: string, minLength: 1 }
                requestId: { type: string, minLength: 1 }
      responses:
        '200':
          description: 举报处理结果
          content:
            application/json:
              schema:
                type: object
                required: [report, duplicate]
                properties:
                  report: { $ref: '#/components/schemas/Report' }
                  duplicate: { type: boolean }
  /admin/audit-log:
    get:
      operationId: listAdminAuditLog
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500 } }
      responses:
        '200':
          description: 平台管理员操作审计
          content:
            application/json:
              schema:
                type: object
                required: [audits]
                properties:
                  audits: { type: array, items: { $ref: '#/components/schemas/AdminAudit' } }
  /admin/score-resets:
    post:
      operationId: resetAllScores
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScoreResetRequest' }
      responses:
        '200':
          description: 新积分周期
          content:
            application/json:
              schema:
                type: object
                required: [epoch, baseScore, audit, duplicate]
                properties:
                  epoch: { type: integer, format: int64 }
                  baseScore: { type: integer, const: 1000 }
                  audit: { $ref: '#/components/schemas/ScoreEpoch' }
                  duplicate: { type: boolean }
        '403': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: rf_session
  parameters:
    RoomID:
      name: roomID
      in: path
      required: true
      schema: { type: string }
  responses:
    Problem:
      description: 统一错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
    User:
      type: object
      required: [id, phone, nickname, permissions, banned, createdAt]
      properties:
        id: { type: string }
        phone: { type: string }
        nickname: { type: string }
        permissions:
          type: object
          additionalProperties: { type: boolean }
        banned: { type: boolean }
        createdAt: { type: string, format: date-time }
    AuthResult:
      type: object
      required: [user, balance]
      properties:
        user: { $ref: '#/components/schemas/User' }
        balance: { type: integer, format: int64 }
    ICEServer:
      type: object
      required: [urls]
      properties:
        urls:
          type: array
          minItems: 1
          items: { type: string }
        username: { type: string }
        credential: { type: string }
    RoomConfig:
      type: object
      additionalProperties: false
      required: [name, maxPlayers, blindPreset, actionSeconds, voiceEnabled, chipDenominations]
      properties:
        name: { type: string, minLength: 1, maxLength: 32 }
        maxPlayers: { type: integer, minimum: 2, maximum: 8 }
        blindPreset: { type: string, enum: ['2/5', '5/10', '10/20'] }
        actionSeconds: { type: integer, enum: [20, 30, 45] }
        voiceEnabled: { type: boolean }
        chipDenominations:
          type: array
          minItems: 5
          maxItems: 8
          uniqueItems: true
          items: { type: integer, enum: [5, 10, 20, 50, 100, 200, 500, 1000] }
...
    TableSnapshot:
      type: object
      required: [roomId, roomCode, roomName, ownerId, version, handNumber, street, pot, board, holeCards, players, allowedChipDenominations, toCall, minimumRaiseBy, maximumRaiseBy, canCheck, canCall, canRaise, canAllIn, actionDeadline, config, messages, ended]
      properties:
        roomId: { type: string }
        roomCode: { type: string }
        roomName: { type: string }
        ownerId: { type: string }
        version: { type: integer, format: int64 }
        handNumber: { type: integer, format: int64 }
        street: { type: string, enum: [waiting, preflop, flop, turn, river, showdown, settled] }
        pot: { type: integer, format: int64 }
        board: { type: array, items: { $ref: '#/components/schemas/PlayingCard' } }
        holeCards: { type: array, items: { $ref: '#/components/schemas/PlayingCard' }, maxItems: 2 }
        players: { type: array, items: { $ref: '#/components/schemas/PlayerSnapshot' } }
        allowedChipDenominations:
          type: array
          items: { type: integer, enum: [5, 10, 20, 50, 100, 200, 500, 1000] }
        toCall: { type: integer, format: int64 }
        minimumRaiseBy: { type: integer, format: int64 }
        maximumRaiseBy: { type: integer, format: int64 }
        canCheck: { type: boolean }
        canCall: { type: boolean }
        canRaise: { type: boolean }
        canAllIn: { type: boolean }
        actionDeadline: { type: string, format: date-time }
        config: { $ref: '#/components/schemas/RoomConfig' }
        messages: { type: array, items: { $ref: '#/components/schemas/SystemMessage' } }
        ended: { type: boolean }
    RoomCommand:
      type: object
      additionalProperties: false
      required: [type, requestId, expectedVersion, payload]
      properties:
        type:
          type: string
          enum: [room.ready, game.start, action.raise, action.call, action.check, action.fold, action.all_in, room.refill, room.leave, room.quick_message, voice.mute, room.remove_player, room.rotate_invite, room.transfer_owner, room.end]
        requestId: { type: string, minLength: 1 }
        expectedVersion: { type: integer, format: int64 }
        payload: { type: object, additionalProperties: true }
    RoomEvent:
      type: object
      required: [type, requestId, roomId, version, sentAt, payload]
      properties:
        type: { type: string }
        requestId: { type: string }
        roomId: { type: string }
        version: { type: integer, format: int64 }
        sentAt: { type: string, format: date-time }
        payload: true
    ScoreAdditionRequest:
      type: object
      additionalProperties: false
      required: [amount, requestId]
      properties:
        amount: { type: integer, minimum: 1, maximum: 1000000000 }
        roomId: { type: string }
        requestId: { type: string, minLength: 1 }
...
    ScoreLedgerEntry:
      type: object
      required: [id, type, amount, balance, note, createdAt, epochId]
      properties:
        id: { type: string }
        type: { type: string, enum: [initial_base, self_add, game_settlement, admin_reset] }
        amount: { type: integer, format: int64 }
        balance: { type: integer, format: int64 }
        roomId: { type: string }
        note: { type: string }
        createdAt: { type: string, format: date-time }
        epochId: { type: integer, format: int64 }
        requestId: { type: string }
...
    Report:
      type: object
      required: [id, reporterId, category, detail, status, createdAt]
      properties:
        id: { type: string }
        reporterId: { type: string }
        roomId: { type: string }
        subjectUserId: { type: string }
        category: { type: string, enum: [conduct, voice, technical, other] }
        detail: { type: string }
        status: { type: string, enum: [open, reviewing, resolved, dismissed] }
        handledBy: { type: string }
        handledAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
...
security:
  - sessionCookie: []